services/identity-service/requirements.txt
django==4.2.7
strawberry-graphql[django]==0.215.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
gunicorn==21.2.0
confluent-kafka
requests